process setup in kernel

+++ include/lib/string.h
@@ -1,5 +1,6 @@
 #include <lib/defs.h>
-void memset(void *data, u8 value, u64 size);
+
+void memset(void *data, int value, u64 size);
 void memcpy(void *dst, const void *src, u64 size);
 
 void do_assert(int, const char *);
